requirements.txt
python-dotenv
torch
torchvision
s3fs
boto3
